Rename of date module
